Update BerserkerHybrid.part
Added the armor proficiency for the class to the heartland feature, emulating how warden does it. Should be balanced. 

P.S., Barbarian skills aren't being marked as class skills, still trying to figure out how to fix/bypass this.

Ignore the previous one. I caused it to create a bug that crashed cbloader.